[Patchset 1] record size

+++ src/profiling/memory/record_reader.h
@@ -41,7 +41,9 @@ class RecordReader {
 
   struct Record {
     std::unique_ptr<uint8_t[]> data;
-    size_t size = 0;
+    // This is not size_t so we can directly copy the received uint64_t
+    // into it.
+    uint64_t size = 0;
   };
 
   ReceiveBuffer BeginReceive();
